chore(cli): add runSentry to cmd
